Context and rules.ts
